+++ packages/core/src/internal/settings/CallbackManager.ts
